Make JWT compileable on Mac OS
yourkarma-JWT depends on SecKeyRawVerify and friends, which are available
on iOS, and according to the iOS header are available on Mac OS X since 10.7,
but which are missing from the Mac SDK headers.

This commit adds extern statements for these missing symbols so that JWT
builds on Mac as well, which works since these symbols are actually
available in the library, they're just not exported.

This bug is tracked with Apple at:
* http://www.openradar.me/25398055
* http://www.openradar.me/22067843